Player now respawns
Death currently uses the world KillZ setting and will respawn the player at a specified spawn point.

Going to switch the code from the World Kill Z to a custom blueprint that allows me to a code a sequence of events before the player actor is destroyed. 

Camera also fades in from black on player spawn.