Tried importing the tower with new normal settings
For some reason, the hive tower keeps coming in with flipped normals. I've tried flipping them myself within 3ds max and letting unreal recalculate the normals, but neither has worked. The placeholder model I made for the hive is actually very similar to the one provided, so I'm using that for now, but it's not unwrapped