Floating Rock Platforms Done
Iteration 2 of the floating platforms is now done. Changed the physical lerp locations to exposed float variables that can be edited from the game scene, as well as added an exposed variable for the amount of time to lerp. 

Started reworking the respawn mechanic so that the character respawned where it left the ground, but that requires some research into getting the ground location before the character changes over to a notGrounded state.