Redid the pause menu when Esc is hit
Hitting escape (and 'z' for for dev purposes) now brings up the pause menu which lets players choose options or exit.

Also reworked the menu controller coding so that it can account for nested menus like the options or quit menus.

Options menu still needs its graphics imported and the functions to be set up.

Main menu needs to be pointed to the new levels and needs it's background level updated.